Add files via upload
Context: 
Malawi’s health budget is currently distributed across districts in a certain fixed proportion, which does not change on a year-on-year basis. It is well understood that this is not based on the expected health expenditure in the district based on the health needs of the population so  an “ideal” distribution of budget has been calculated towards which the Ministry of Health wishes to move over a certain period of time (in order to avoid large shocks, either positive or negative.) 

Task: 
This python code runs a simulation to estimate the number of years it would take to arrive at the new distribution assuming a certain rate of growth of the total budget available and also assuming that no district experiences a reduction in its absolute budget.